test/e2e: move funcs from test/e2e/pod to other folders
